+++ src/lib.rs
@@ -53,6 +53,7 @@ pub use reqwest;
 
 pub mod cache;
 pub mod config;
+pub mod engine;
 pub mod http;
 pub mod metrics;
 pub mod operation;
@@ -64,6 +65,5 @@ pub mod rta;
 pub mod slurm;
 pub mod store;
 pub mod utils;
-pub mod validation;
 pub mod validity;
